Properly handle the config.json
